Remove geographic based solar panel placement.
