+++ src/algorithm/__tests__/solarPanelPlacement.spec.ts
@@ -2,50 +2,12 @@ import { describe, expect, test } from 'vitest'
 import * as fs from 'fs'
 import * as path from 'path'
 import {
-  solarPanelPlacementAlgorithm,
-  solarPanelPlacement,
-  extractCentroids,
-} from '../solarPanelPlacement'
-import {
   solarPanelPlacement as solarPanelPlacementMatrix,
   Matrix,
   rearrangeMatrix,
 } from '../solarPanelPlacementMatrix'
-import type { FeatureCollection, Polygon, Properties } from '@turf/turf'
 
 describe('solar panel placement', () => {
-  test('test solar panel algorithm', () => {
-    // Read GeoJSON file
-    const fileContents = fs.readFileSync(
-      path.join(__dirname, 'grids.geojson'),
-      {
-        encoding: 'utf-8',
-      }
-    )
-
-    const grid: FeatureCollection<Polygon, Properties> =
-      JSON.parse(fileContents)
-
-    const horizontalSolarPanels = solarPanelPlacementAlgorithm(grid, true, true)
-    expect(horizontalSolarPanels.features.length).greaterThanOrEqual(9)
-
-    const verticalSolarPanels = solarPanelPlacementAlgorithm(grid, false, true)
-    // In this test case, the vertical placement is not optimum, only 8
-    expect(verticalSolarPanels.features.length).greaterThanOrEqual(8)
-
-    const solarPanelsPlacement = solarPanelPlacement(grid)
-    expect(solarPanelsPlacement.solarPanels.features.length).greaterThanOrEqual(
-      9
-    )
-    expect(solarPanelsPlacement.orientation).equal('horizontal')
-
-    const solarPanelCentroids = extractCentroids(
-      solarPanelsPlacement.solarPanels,
-      true
-    )
-    expect(solarPanelCentroids.features.length).equal(9)
-  })
-
   test('test solar panel algorithm from matrix', () => {
     // Read GeoJSON file
     const fileContents = fs.readFileSync(
